@@ -17,11 +17,11 @@
     }
   ],
   "require": {
-    "php": ">=5.3.3",
-    "guzzle/guzzle": "~3.5"
+    "php": ">=5.4",
+    "guzzlehttp/guzzle-services": "~0.3"
   },
   "require-dev": {
-    "phpunit/phpunit": "~3.7"
+    "phpunit/phpunit": "~4.1"
   },
   "autoload": {
     "psr-0": {
