@@ -33,7 +33,7 @@ class ErrorHandlerListenerTest extends PHPUnit_Framework_TestCase
         $errorHandler = new ErrorHandlerListener();
         $event        = new Event();
 
-        $response = $this->getMock('Guzzle\Http\Message\Response', array(), array(), '', false);
+        $response = $this->getMock('Guzzle\Http\Message\Response', [], [], '', false);
         $command  = $this->getMock('Guzzle\Service\Command\CommandInterface');
 
         $command->expects($this->once())->method('getResponse')->will($this->returnValue($response));
@@ -46,95 +46,95 @@ class ErrorHandlerListenerTest extends PHPUnit_Framework_TestCase
 
     public function errorProvider()
     {
-        return array(
-            array('Absplit_UnknownError'),
-            array('Absplit_UnknownSplitTest'),
-            array('Absplit_UnknownTestType'),
-            array('Absplit_UnknownWaitUnit'),
-            array('Absplit_UnknownWinnerType'),
-            array('Absplit_WinnerNotSelected'),
-            array('Avesta_Db_Exception'),
-            array('Campaign_BounceMissing'),
-            array('Campaign_DoesNotExist'),
-            array('Campaign_InvalidAbsplit'),
-            array('Campaign_InvalidAuto'),
-            array('Campaign_InvalidContent'),
-            array('Campaign_InvalidOption'),
-            array('Campaign_InvalidRss'),
-            array('Campaign_InvalidSegment'),
-            array('Campaign_InvalidStatus'),
-            array('Campaign_InvalidTemplate'),
-            array('Campaign_NotSaved'),
-            array('Campaign_StatsNotAvailable'),
-            array('Email_AlreadySubscribed'),
-            array('Email_AlreadyUnsubscribed'),
-            array('Email_NotExists'),
-            array('Email_NotSubscribed'),
-            array('Invalid_Analytics'),
-            array('Invalid_ApiKey'),
-            array('Invalid_AppKey'),
-            array('Invalid_DateTime'),
-            array('Invalid_EcommOrder'),
-            array('Invalid_Email'),
-            array('Invalid_Folder'),
-            array('Invalid_IP'),
-            array('Invalid_Options'),
-            array('Invalid_PagingLimit'),
-            array('Invalid_PagingStart'),
-            array('Invalid_SendType'),
-            array('Invalid_Template'),
-            array('Invalid_TrackingOptions'),
-            array('Invalid_URL'),
-            array('List_AlreadySubscribed'),
-            array('List_CannotRemoveEmailMerge'),
-            array('List_DoesNotExist'),
-            array('List_InvalidBounceMember'),
-            array('List_InvalidImport'),
-            array('List_InvalidInterestFieldType'),
-            array('List_InvalidInterestGroup'),
-            array('List_InvalidMergeField'),
-            array('List_InvalidOption'),
-            array('List_InvalidUnsubMember'),
-            array('List_MergeFieldRequired'),
-            array('List_Merge_InvalidMergeID'),
-            array('List_NotSubscribed'),
-            array('List_TooManyInterestGroups'),
-            array('List_TooManyMergeFields'),
-            array('MC_ContentImport_InvalidArchive'),
-            array('MC_InvalidPayment'),
-            array('MC_PastedList_Duplicate'),
-            array('MC_PastedList_InvalidImport'),
-            array('MC_SearchException'),
-            array('Max_Size_Reached'),
-            array('Module_Unknown'),
-            array('MonthlyPlan_Unknown'),
-            array('Order_TypeUnknown'),
-            array('PDOException'),
-            array('Parse_Exception'),
-            array('Request_TimedOut'),
-            array('ServerError_InvalidParameters'),
-            array('ServerError_MethodUnknown'),
-            array('Too_Many_Connections'),
-            array('Unknown_Exception'),
-            array('User_CannotSendCampaign'),
-            array('User_Disabled'),
-            array('User_DoesExist'),
-            array('User_DoesNotExist'),
-            array('User_InvalidAction'),
-            array('User_InvalidRole'),
-            array('User_MissingEmail'),
-            array('User_MissingModuleOutbox'),
-            array('User_ModuleAlreadyPurchased'),
-            array('User_ModuleNotPurchased'),
-            array('User_NotApproved'),
-            array('User_NotEnoughCredit'),
-            array('User_UnderMaintenance'),
-            array('User_Unknown'),
-            array('ValidationError'),
-            array('XML_RPC2_Exception'),
-            array('XML_RPC2_FaultException'),
-            array('Zend_Uri_Exception')
-        );
+        return [
+            ['Absplit_UnknownError'],
+            ['Absplit_UnknownSplitTest'],
+            ['Absplit_UnknownTestType'],
+            ['Absplit_UnknownWaitUnit'],
+            ['Absplit_UnknownWinnerType'],
+            ['Absplit_WinnerNotSelected'],
+            ['Avesta_Db_Exception'],
+            ['Campaign_BounceMissing'],
+            ['Campaign_DoesNotExist'],
+            ['Campaign_InvalidAbsplit'],
+            ['Campaign_InvalidAuto'],
+            ['Campaign_InvalidContent'],
+            ['Campaign_InvalidOption'],
+            ['Campaign_InvalidRss'],
+            ['Campaign_InvalidSegment'],
+            ['Campaign_InvalidStatus'],
+            ['Campaign_InvalidTemplate'],
+            ['Campaign_NotSaved'],
+            ['Campaign_StatsNotAvailable'],
+            ['Email_AlreadySubscribed'],
+            ['Email_AlreadyUnsubscribed'],
+            ['Email_NotExists'],
+            ['Email_NotSubscribed'],
+            ['Invalid_Analytics'],
+            ['Invalid_ApiKey'],
+            ['Invalid_AppKey'],
+            ['Invalid_DateTime'],
+            ['Invalid_EcommOrder'],
+            ['Invalid_Email'],
+            ['Invalid_Folder'],
+            ['Invalid_IP'],
+            ['Invalid_Options'],
+            ['Invalid_PagingLimit'],
+            ['Invalid_PagingStart'],
+            ['Invalid_SendType'],
+            ['Invalid_Template'],
+            ['Invalid_TrackingOptions'],
+            ['Invalid_URL'],
+            ['List_AlreadySubscribed'],
+            ['List_CannotRemoveEmailMerge'],
+            ['List_DoesNotExist'],
+            ['List_InvalidBounceMember'],
+            ['List_InvalidImport'],
+            ['List_InvalidInterestFieldType'],
+            ['List_InvalidInterestGroup'],
+            ['List_InvalidMergeField'],
+            ['List_InvalidOption'],
+            ['List_InvalidUnsubMember'],
+            ['List_MergeFieldRequired'],
+            ['List_Merge_InvalidMergeID'],
+            ['List_NotSubscribed'],
+            ['List_TooManyInterestGroups'],
+            ['List_TooManyMergeFields'],
+            ['MC_ContentImport_InvalidArchive'],
+            ['MC_InvalidPayment'],
+            ['MC_PastedList_Duplicate'],
+            ['MC_PastedList_InvalidImport'],
+            ['MC_SearchException'],
+            ['Max_Size_Reached'],
+            ['Module_Unknown'],
+            ['MonthlyPlan_Unknown'],
+            ['Order_TypeUnknown'],
+            ['PDOException'],
+            ['Parse_Exception'],
+            ['Request_TimedOut'],
+            ['ServerError_InvalidParameters'],
+            ['ServerError_MethodUnknown'],
+            ['Too_Many_Connections'],
+            ['Unknown_Exception'],
+            ['User_CannotSendCampaign'],
+            ['User_Disabled'],
+            ['User_DoesExist'],
+            ['User_DoesNotExist'],
+            ['User_InvalidAction'],
+            ['User_InvalidRole'],
+            ['User_MissingEmail'],
+            ['User_MissingModuleOutbox'],
+            ['User_ModuleAlreadyPurchased'],
+            ['User_ModuleNotPurchased'],
+            ['User_NotApproved'],
+            ['User_NotEnoughCredit'],
+            ['User_UnderMaintenance'],
+            ['User_Unknown'],
+            ['ValidationError'],
+            ['XML_RPC2_Exception'],
+            ['XML_RPC2_FaultException'],
+            ['Zend_Uri_Exception']
+        ];
     }
 
     /**
@@ -145,14 +145,14 @@ class ErrorHandlerListenerTest extends PHPUnit_Framework_TestCase
         $errorHandler = new ErrorHandlerListener();
         $event        = new Event();
 
-        $response = $this->getMock('Guzzle\Http\Message\Response', array(), array(), '', false);
+        $response = $this->getMock('Guzzle\Http\Message\Response', [], [], '', false);
         $command  = $this->getMock('Guzzle\Service\Command\CommandInterface');
 
-        $responseData = array(
+        $responseData = [
             'name'  => $errorName,
             'code'  => 100, // dummy code just for test
             'error' => 'Error message' // dummy error message just for test
-        );
+        ];
 
         $command->expects($this->once())->method('getResponse')->will($this->returnValue($response));
         $command->expects($this->once())->method('toArray')->will($this->returnValue($responseData));
